feat: add and implement the Post Service. fix bugs with parameters, add additional fields in dtos.

+++ src/main/java/org/alphatrack/screensociety/dto/request/filters/CommentFilterOptions.java
@@ -9,15 +9,15 @@ public class CommentFilterOptions {
 
     private final Optional<String> content;
     private final Optional<String> authorUsername;
-    private final Optional<Integer> minLength;
-    private final Optional<Integer> maxLength;
+    private final Optional<Long> minLength;
+    private final Optional<Long> maxLength;
     private final Optional<String> sortBy;
     private final Optional<String> sortOrder;
 
     public CommentFilterOptions(String content,
                                 String authorUsername,
-                                Integer minLength,
-                                Integer maxLength,
+                                Long minLength,
+                                Long maxLength,
                                 String sortBy,
                                 String sortOrder) {
         this.content = Optional.ofNullable(content);
